Code cleanup and fixing compilation warnings related to nullability
